modules/mod-001-python-fundamentals/exercise-01/sentiment-classifier/requirements.txt
# Production Dependencies
# All versions pinned for reproducibility

# Deep Learning
torch==2.1.0
transformers==4.35.0

# Data Processing
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.2

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
tqdm==4.66.1
